Update hash, allow offline installation.

+++ install.sh
@@ -4,7 +4,9 @@
 URL="https://github.com/cdr/code-server/releases/download/v3.8.0/code-server_3.8.0_amd64.deb"
 CODE_SERVER_DEB_PATH="/tmp/code-server.deb"
 rm $CODE_SERVER_DEB_PATH 2>/dev/null
-curl -fL $URL -o $CODE_SERVER_DEB_PATH
+if [ ! -f $CODE_SERVER_DEB_PATH ]; then
+    curl -fL $URL -o $CODE_SERVER_DEB_PATH
+fi
 
 # Verify integrity.
 SHA256SUM=$(sha256sum $CODE_SERVER_DEB_PATH | awk '{ print $1 }')
@@ -35,15 +37,16 @@ sudo systemctl enable --now code-server@liman
 URL="https://github.com/mertcelen/liman-gelistirme-modulu/releases/download/latest/liman-gelistirme-modulu.deb"
 MODULE_DEB_PATH="/tmp/liman-gelistirme-modulu.deb"
 rm $MODULE_DEB_PATH 2>/dev/null
-curl -fL $URL -o $MODULE_DEB_PATH
-
+if [ ! -f $MODULE_DEB_PATH ]; then
+    curl -fL $URL -o $MODULE_DEB_PATH
+fi
 # Install Deb File
 sudo apt install $MODULE_DEB_PATH -yyq
 
 # Verify integrity.
 SHA256SUM=$(sha256sum $MODULE_DEB_PATH | awk '{ print $1 }')
 
-if [[ $SHA256SUM != "" ]]; then
+if [[ $SHA256SUM != "1d4778d1ef01b3a5891755e7776cc347b0010f94edef9a4a9c26917faed5f914" ]]; then
   echo "File download failed, please download $URL and move it as $MODULE_DEB_PATH"
   exit 1
 fi
